split parser generic declarations
