+++ src/parser/declarations/generic.rs
@@ -1,0 +1,113 @@
+use super::*;
+use crate::parser::keywords::ParserBehaviorKeyword;
+
+impl Parser {
+    pub(in crate::parser) fn parse_generic_declaration(
+        &mut self,
+        name: String,
+        type_params: Vec<TypeParam>,
+        public: bool,
+        name_span: Span,
+    ) -> Result<Declaration, CompileError> {
+        self.skip_newlines();
+        match self.peek() {
+            Token::Assign => {
+                // Generic function: `name<T> = (params) ret { body }`
+                self.parse_function_def(name, type_params, public, name_span)
+            }
+            Token::Dot => {
+                // Generic receiver method: `Type<T>.method = (self: Type<T>) ...`
+                self.advance();
+                let (method_name, _method_span) = self.expect_identifier()?;
+                self.skip_newlines();
+
+                if let Ok(keyword) = method_name.parse::<TypeDeclarationKeyword>() {
+                    if matches!(keyword, TypeDeclarationKeyword::Impl) {
+                        return self.parse_impl_block_with_type_params(
+                            name,
+                            type_params,
+                            name_span,
+                        );
+                    }
+                    if matches!(keyword, TypeDeclarationKeyword::Implements) {
+                        return self.parse_behavior_impl_block_with_type_params(
+                            name,
+                            type_params,
+                            name_span,
+                        );
+                    }
+                    return self.reject_gated_generic_association_target(keyword, name_span);
+                }
+
+                let mut all_type_params = type_params;
+                if matches!(self.peek(), Token::Lt) {
+                    all_type_params.extend(self.parse_type_params()?);
+                }
+
+                self.skip_newlines();
+                self.expect(&Token::Assign)?;
+                self.skip_newlines();
+
+                let (params, return_type, body) = self.parse_function_signature_and_body()?;
+                let span = name_span.merge(body.span());
+                Ok(Declaration::Method {
+                    type_name: name,
+                    method_name,
+                    type_params: all_type_params,
+                    params,
+                    return_type,
+                    body,
+                    public,
+                    span,
+                })
+            }
+            Token::Colon if self.is_struct_def() => {
+                self.parse_struct_def_with_params(name, type_params, public, name_span)
+            }
+            Token::Colon
+                if self
+                    .colon_is_followed_by_identifier(ParserBehaviorKeyword::Behavior.as_str()) =>
+            {
+                self.parse_behavior_def(name, type_params, public, name_span)
+            }
+            Token::Colon => self.parse_enum_def_with_params(name, type_params, public, name_span),
+            _ => Err(CompileError::Syntax(
+                "expected '=' or ':' after generic type parameters".to_string(),
+                Some(self.peek_span()),
+            )),
+        }
+    }
+
+    fn reject_gated_generic_association_target<T>(
+        &mut self,
+        keyword: TypeDeclarationKeyword,
+        name_span: Span,
+    ) -> Result<T, CompileError> {
+        let span = self.gated_association_call_span(name_span);
+        Err(CompileError::Syntax(
+            format!(
+                "generic association target `Type<T>.{keyword}` is gated; use non-generic `{keyword}` associations or keep the generic behavior target deferred to docs/V1_SPEC.md"
+            ),
+            Some(span),
+        ))
+    }
+
+    fn gated_association_call_span(&mut self, name_span: Span) -> Span {
+        if !matches!(self.peek(), Token::LParen) {
+            return name_span.merge(self.prev_span());
+        }
+
+        self.advance();
+        let behavior_span = self.expect_identifier().map(|(_, span)| span).ok();
+        if matches!(self.peek(), Token::Lt) {
+            let _ = self.parse_type_arg_list();
+        }
+        self.skip_newlines();
+        match self.expect(&Token::RParen) {
+            Ok(end) => name_span.merge(end),
+            Err(_) => behavior_span
+                .map(|span| name_span.merge(span))
+                .unwrap_or_else(|| name_span.merge(self.prev_span())),
+        }
+    }
+}
